+++ CHANGELOG.md
@@ -9,6 +9,7 @@ v0.3.0 - unreleased
 - `ff::Direction::from_radians(f32)` and `ff::Direction::from_vector([f32; 2])`
 - `Gilrs::gamepads(&self)` which returns iterator over all connected gamepads
 - `GamepadState` now implements `is_btn_pressed(Button)` and `axis_val(Axis)`
+- `Gilrs` now implements `Index`and `IndexMut`
 
 ### Changed